# app/static/swagger.yaml
swagger: "2.0"
...
info:
  title: "Mechanic Database"
  description: "API for managing a mechanic shop system, including customers, service tickets, mechanics, and inventory."
  version: "1.0.0"

# API Host and Schemes
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# Security Definitions
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"

# API Endpoints
paths:
...
  /service-tickets:
    get:
      tags:
        - "service-tickets"
      summary: "List all service tickets"
      description: "Retrieve a list of all service tickets."
      responses:
        200:
          description: "Service tickets retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTicketResponse"
    post:
      tags:
        - "service-tickets"
      summary: "Create a service ticket"
      description: "Create a new service ticket (requires customer authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Service ticket information"
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:
        201:
          description: "Service ticket created successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
  /service-tickets/my-tickets:
    get:
      tags:
        - "service-tickets"
      summary: "List customer's service tickets"
      description: "Retrieve all service tickets for the authenticated customer."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer tickets retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTicketResponse"
  /service-tickets/{ticket_id}:
    get:
      tags:
        - "service-tickets"
      summary: "Get a service ticket"
      description: "Retrieve details of a specific service ticket by ID."
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
      responses:
        200:
          description: "Service ticket retrieved successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
    put:
      tags:
        - "service-tickets"
      summary: "Update a service ticket"
      description: "Update an existing service ticket (requires customer authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
        - in: "body"
          name: "body"
          description: "Updated service ticket information"
          required: true
          schema:
            $ref: "#/definitions/UpdateServiceTicketPayload"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
    delete:
      tags:
        - "service-tickets"
      summary: "Delete a service ticket"
      description: "Remove a service ticket from the system."
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
      responses:
        200:
          description: "Service ticket deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"
  /service-tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
        - "service-tickets"
      summary: "Assign mechanic to service ticket"
      description: "Assign a mechanic to a specific service ticket."
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic to assign"
      responses:
        200:
          description: "Mechanic assigned successfully"
  /service-tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
        - "service-tickets"
      summary: "Remove mechanic from service ticket"
      description: "Remove a mechanic from a specific service ticket."
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic to remove"
      responses:
        200:
          description: "Mechanic removed successfully"
  /service-tickets/{ticket_id}/edit-mechanics:
    put:
      tags:
        - "service-tickets"
      summary: "Edit mechanics on service ticket"
      description: "Add or remove multiple mechanics from a service ticket (requires customer authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
        - in: "body"
          name: "body"
          description: "Mechanics to add or remove"
          required: true
          schema:
            $ref: "#/definitions/EditMechanicsPayload"
      responses:
        200:
          description: "Mechanics updated successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
  /service-tickets/{ticket_id}/labor:
    post:
      tags:
        - "service-tickets"
      summary: "Log labor hours"
      description: "Log labor hours for a mechanic on a service ticket (requires mechanic authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
        - in: "body"
          name: "body"
          description: "Labor log information"
          required: true
          schema:
            $ref: "#/definitions/LaborLogPayload"
      responses:
        201:
          description: "Labor hours logged successfully"
          schema:
            $ref: "#/definitions/LaborLogResponse"
  /service-tickets/labor/{labor_log_id}:
    put:
      tags:
        - "service-tickets"
      summary: "Update labor log"
      description: "Update labor hours logged by a mechanic (requires mechanic authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "labor_log_id"
          type: "integer"
          required: true
          description: "ID of the labor log"
        - in: "body"
          name: "body"
          description: "Updated labor hours"
          required: true
          schema:
            $ref: "#/definitions/UpdateLaborLogPayload"
      responses:
        200:
          description: "Labor log updated successfully"
          schema:
            $ref: "#/definitions/LaborLogResponse"
    delete:
      tags:
        - "service-tickets"
      summary: "Delete labor log"
      description: "Remove a labor log entry."
      parameters:
        - in: "path"
          name: "labor_log_id"
          type: "integer"
          required: true
          description: "ID of the labor log"
      responses:
        200:
          description: "Labor log deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"

  # Mechanic Endpoints
...
  /mechanics/{mechanic_id}/service_tickets:
    get:
      tags:
        - "mechanics"
      summary: "List mechanic's service tickets"
      description: "Retrieve all service tickets assigned to a specific mechanic (requires mechanic authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic"
      responses:
        200:
          description: "Mechanic's service tickets retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTicketResponse"
...
definitions:
...
  CreateServiceTicketPayload:
    type: "object"
    properties:
      description:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      VIN:
        type: "string"
    required:
      - description
      - service_date
      - VIN
  ServiceTicketResponse:
    type: "object"
    properties:
      customer_id:
        type: "integer"
      description:
        type: "string"
      labor_logs:
        type: "array"
        items:
          type: "object"
      mechanics:
        type: "array"
        items:
          type: "object"
          properties:
            email:
              type: "string"
            id:
              type: "integer"
            name:
              type: "string"
      service_date:
        type: "string"
      ticket_id:
        type: "integer"
      VIN:
        type: "string"
  UpdateServiceTicketPayload:
    type: "object"
    properties:
      description:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      VIN:
        type: "string"
  LaborLogPayload:
    type: "object"
    properties:
      hours_worked:
        type: "number"
      mechanic_id:
        type: "integer"
    required:
      - hours_worked
      - mechanic_id
  LaborLogResponse:
    type: "object"
    properties:
      date_logged:
        type: "string"
      hours_worked:
        type: "number"
      id:
        type: "integer"
      mechanic_id:
        type: "integer"
      ticket_id:
        type: "integer"
  UpdateLaborLogPayload:
    type: "object"
    properties:
      hours_worked:
        type: "number"
    required:
      - hours_worked
  EditMechanicsPayload:
    type: "object"
    properties:
      add_mechanic_ids:
        type: "array"
        items:
          type: "integer"
      remove_mechanic_ids:
        type: "array"
        items:
          type: "integer"

  # Mechanic Schemas
...
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
...